Add subdirectory 'team' under 'components'
