+++ front/components/team/TeamDrawing.module.css
